forgot casting data type on teamstat / teamstat request packet

+++ tools/DedicatedClient/main.cpp
@@ -416,7 +416,7 @@ bool UpdateClientNet()
 			case NETMSG_TEAMSTAT:
 			{
 				int team=inbuf[1];
-				int frameCount = inbuf[2];
+				int frameCount = static_cast<unsigned int>(inbuf[2]);
 				if( team > gameSetup->teamStartingData.size() ){
 					logOutput.Print("Got invalid team num %i in teamstat msg",team);
 					break;
